langchain_genai/traditional_cusine_mini_project/requirements.txt
langchain
langchain_core
langchain_openai
streamlit
python-dotenv